Fix auto-scroll of the menu
Auto-scroll of the menu will first scroll the TOC into view if needed.

If no TOC found, it will scroll the active menu item if needed.

We are using scroll-into-view-if-needed to do the job for us, as it allows to set boundary avoiding scrolling the main content of the page.